Gui: split selection implementations
